requirements.txt
tensorflow
opencv-python
numpy
pillow
matplotlib
seaborn
scikit-learn
streamlit